requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-multipart>=0.0.9
httpx>=0.27.0
apscheduler>=3.10.4
playwright>=1.44.0
pypdf>=4.2.0
pystray>=0.19.5
Pillow>=10.3.0
pyinstaller>=6.6.0
